+++ app/[locale]/page.tsx
@@ -1,8 +1,11 @@
 import Hero from "../components/Hero";
+// import ConsentAwareComponent from "../components/ConsentAwareComponent";
+
 
 export default function Home() {
   return (
     <main>
+      {/* <ConsentAwareComponent /> */}
       <Hero />
     </main>
   )