+++ next.config.js
@@ -1,5 +1,6 @@
 const withNextIntl = require('next-intl/plugin')();
  
 module.exports = withNextIntl({
+  transpilePackages: ['@mep-agency/next-iubenda'],
   // Other Next.js configuration ...
 });