+++ CHANGELOG.md
@@ -1,6 +1,25 @@
+2024-05-02␣␣Stefano Renzetti␣␣<renzetti.graphics@gmail.com>
+
+* Added the `Privacy Policy` button in `Footer.tsx`.
+
+* Added `min-width-full` class to hero's image for Firefox.
+
+* Disabled the Iubenda's cookie solution. To be FIX. 
+
+2024-02-02␣␣Stefano Renzetti␣␣<renzetti.graphics@gmail.com>
+
+* Created `ConsentAwareComponent.tsx` to use `useIubenda()` hooks.
+
+* Installed Iubenda library "Next-Iubenda" with
+  `yarn add --dev @mep-agency/next-iubenda` for Cookies.
+
+* TO SEARCH: GITHUB PAGES SUPPORTA IL MIDDLEWARE
+
+* Created `IubendaPrivacyPolicy` for display Iubenda's privacy.
+
 2024-18-01␣␣Stefano Renzetti␣␣<renzetti.graphics@gmail.com>
 
-* Installed i18n library "next-intl with `yarn add next-intl`.
+* Installed i18n library "next-intl" with `yarn add next-intl`.
 
 * Created the config file `i18n.tsx`.